requirements.txt
# torch
torchvision
tqdm
matplotlib
numpy
seaborn
torchaudio
librosa
parse-utils
ipywidgets
transformers
datasets